+++ src/containers/report/ActivitySummarySection.tsx
