+++ src/containers/report/CompetencyStatsSection.tsx
