+++ src/containers/report/ExperienceHighlightsSection.tsx
