+++ src/containers/report/NarrativeSummarySection.tsx
