+++ src/containers/report/NextFocusPointSection.tsx
