+++ src/containers/report/PatternSection.tsx
