+++ src/containers/report/StrengthsSection.tsx
